Bug 1466178 - Reorganize directories for the detail-panel
