+++ ui/entry-index.js
@@ -51,11 +51,11 @@ import './js/controllers/tcjobactions';
 import './plugins/tabs';
 import './plugins/controller';
 import './plugins/pinboard';
-import './details-panel/JobDetailsPane';
-import './details-panel/JobDetailsTab';
-import './details-panel/FailureSummaryTab';
-import './details-panel/AutoclassifyTab';
-import './details-panel/AnnotationsTab';
-import './details-panel/SimilarJobsTab';
-import './details-panel/PerformanceTab';
+import './job-view/details-panel/summary-panel/SummaryPanel';
+import './job-view/details-panel/tabs-panel/JobDetailsTab';
+import './job-view/details-panel/tabs-panel/FailureSummaryTab';
+import './job-view/details-panel/tabs-panel/AutoclassifyTab';
+import './job-view/details-panel/tabs-panel/AnnotationsTab';
+import './job-view/details-panel/tabs-panel/SimilarJobsTab';
+import './job-view/details-panel/tabs-panel/PerformanceTab';
 import './js/filters';
