+++ ui/job-view/details-panel/tabs-panel/JobDetailsTab.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { react2angular } from 'react2angular/index.es2015';
 
-import treeherder from '../js/treeherder';
-import { getPerfAnalysisUrl, getWptUrl } from "../helpers/url";
+import treeherder from '../../../js/treeherder';
+import { getPerfAnalysisUrl, getWptUrl } from '../../../helpers/url';
 
 export default class JobDetailsTab extends React.PureComponent {
   render() {
