+++ ui/job-view/details-panel/tabs-panel/PerformanceTab.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { react2angular } from 'react2angular/index.es2015';
 
-import treeherder from '../js/treeherder';
-import { getCompareChooserUrl } from '../helpers/url';
+import treeherder from '../../../js/treeherder';
+import { getCompareChooserUrl } from '../../../helpers/url';
 
 class PerformanceTab extends React.PureComponent {
