+++ ui/job-view/details-panel/tabs-panel/autoclassify/ErrorLineModel.js
