+++ ui/job-view/details-panel/tabs-panel/failureSummary/ErrorsList.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export default function ErrorsList(props) {
+  const errorListItem = props.errors.map((error, key) => (
+    <li
+      key={key} // eslint-disable-line react/no-array-index-key
+    >{error.name} : {error.result}.
+      <a
+        title="Open in Log Viewer"
+        target="_blank"
+        rel="noopener"
+        href={error.lvURL}
+      ><span className="ml-1">View log</span></a>
+    </li>
+  ));
+
+  return (
+    <li>
+      No Bug Suggestions Available.<br />
+      <span className="font-weight-bold">Unsuccessful Execution Steps</span>
+      <ul>{errorListItem}</ul>
+    </li>
+  );
+}
+
+ErrorsList.propTypes = {
+  errors: PropTypes.array.isRequired,
+};
