+++ ui/js/constants.js
@@ -371,3 +371,5 @@ export const phAlertStatusMap = {
 };
 
 export const phCompareBaseLineDefaultTimeRange = 86400 * 2;
+
+export const thBugSuggestionLimit = 20;
