+++ ui/plugins/pluginpanel.html
@@ -138,8 +138,8 @@
       </nav>
     </div>
     <div id="job-details-pane">
-      <job-details-pane
-        name="JobDetailsPane"
+      <summary-panel
+        name="SummaryPanel"
         classification="latestClassification"
         job="job"
         bugs="bugs"
